+++ src/skmatter/decomposition/_pcovc.py
@@ -91,13 +91,18 @@ def check_cl_fit(classifier, X, y):
         fitted_classifier._validate_data(X, y, reset=False, multi_output=True)
 
         # Check compatibility with y
-        if fitted_classifier.coef_.ndim != y.ndim:
+
+        # changed from if fitted_classifier.coef_.ndim != y.ndim:
+        # dimension of classifier coefficients is always 2, hence we don't need to check 
+        # for match with Y
+        if fitted_classifier.coef_.shape[1] != X.shape[1]:
             raise ValueError(
-                "The classifier coefficients have a dimension incompatible "
-                "with the supplied target space. "
-                "The coefficients have dimension %d and the targets "
-                "have dimension %d" % (fitted_classifier.coef_.ndim, y.ndim)
+                "The classifier coefficients have a shape incompatible "
+                "with the supplied feature space. "
+                "The coefficients have shape %d and the features "
+                "have shape %d" % (fitted_classifier.coef_.shape, X.shape)
             )
+        # LogisticRegression does not support multioutput, but RidgeClassifier does
         elif y.ndim == 2:
             if fitted_classifier.coef_.shape[0] != y.shape[1]:
                 raise ValueError(
